backend/requirements.txt
flask
flask-cors
sqlalchemy
requests
beautifulsoup4
python-dotenv
pandas
selenium
webdriver-manager
sqlalchemy-utils
flask-sqlalchemy